+++ common/src/main/kotlin/com/cobblemon/mod/common/api/pokemon/experience/ExperienceCalculator.kt
@@ -9,6 +9,7 @@
 package com.cobblemon.mod.common.api.pokemon.experience
 
 import com.cobblemon.mod.common.battles.pokemon.BattlePokemon
+import com.cobblemon.mod.common.pokemon.evolution.requirements.LevelRequirement
 import kotlin.math.pow
 
 interface ExperienceCalculator {
@@ -30,8 +31,11 @@ object StandardExperienceCalculator : ExperienceCalculator {
         val pointPowerMultiplier = 1.0
         val participatedMultiplier = if (opponentPokemon in battlePokemon.facedOpponents) 1.0 else 2.0 // Implement after Exp Share etc
         val originalTrainerMultiplier = 1.0
-        val evolutionMultiplier = 1.0 // implement after evolutions, it is 'v' on bulbapedia's equation
-
+        val evolutionMultiplier =
+            if (battlePokemon.effectedPokemon.evolutionProxy.server().any { evolution ->
+                val requirements = evolution.requirements.asSequence()
+                requirements.any { it is LevelRequirement } && requirements.all { it.check(battlePokemon.effectedPokemon) }
+            }) 1.2 else 1.0
         val term1 = (baseExp * level * affectionMultiplier * evolutionMultiplier) / (5 * participatedMultiplier)
         val term2 = ((2.0 * level + 10) / (level + levelVictor + 10)).pow(2.5)
         val term3 = originalTrainerMultiplier * luckyEggMultiplier * pointPowerMultiplier