+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fixed optional aspects not saving and loading properly.
 - Fixed issues of invisible Pokémon in GUIs on some Macs.
 - Fixed a soft-duplicate that could occur when a party Pokémon is pushed through a Nether Portal.
+- Fixed missing multiplier on the experience calculator when a Pokémon is at or past the level where it would be able to evolve, but it has not.
 
 ## [1.2.0 - The Customization Update (January 1st, 2023)](#1-2-0)
 ### Additions